refactor: use pane layout component in pages
