+++ packages/hoppscotch-app/components/realtime/Socketio.vue
@@ -1,103 +1,14 @@
 <template>
-  <Splitpanes
-    class="smart-splitter"
-    :rtl="SIDEBAR_ON_LEFT && mdAndLarger"
-    :class="{
-      '!flex-row-reverse': SIDEBAR_ON_LEFT && mdAndLarger,
-    }"
-    :horizontal="!mdAndLarger"
-  >
-    <Pane
-      size="75"
-      min-size="65"
-      class="hide-scrollbar !overflow-auto flex flex-col"
-    >
-      <Splitpanes class="smart-splitter" :horizontal="COLUMN_LAYOUT">
-        <Pane
-          :size="COLUMN_LAYOUT ? 45 : 50"
-          class="hide-scrollbar !overflow-auto flex flex-col"
-        >
-          <div
-            class="sticky top-0 z-10 flex flex-shrink-0 p-4 overflow-x-auto space-x-2 bg-primary hide-scrollbar"
-          >
-            <div class="inline-flex flex-1 space-x-2">
-              <div class="flex flex-1">
-                <label for="client-version">
-                  <tippy
-                    ref="versionOptions"
-                    interactive
-                    trigger="click"
-                    theme="popover"
-                    arrow
-                  >
-                    <template #trigger>
-                      <span class="select-wrapper">
-                        <input
-                          id="client-version"
-                          v-tippy="{ theme: 'tooltip' }"
-                          title="socket.io-client version"
-                          class="flex px-4 py-2 font-semibold border rounded-l cursor-pointer bg-primaryLight border-divider text-secondaryDark w-26"
-                          :value="`Client ${clientVersion}`"
-                          readonly
-                          :disabled="connectionState"
-                        />
-                      </span>
-                    </template>
-                    <div class="flex flex-col" role="menu">
-                      <SmartItem
-                        v-for="(_, version) in socketIoClients"
-                        :key="`client-${version}`"
-                        :label="`Client ${version}`"
-                        @click.native="onSelectVersion(version)"
-                      />
-                    </div>
-                  </tippy>
-                </label>
-                <input
-                  id="socketio-url"
-                  v-model="url"
-                  type="url"
-                  autocomplete="off"
-                  spellcheck="false"
-                  :class="{ error: !urlValid }"
-                  class="flex flex-1 w-full px-4 py-2 border bg-primaryLight border-divider text-secondaryDark"
-                  :placeholder="$t('socketio.url')"
-                  :disabled="connectionState"
-                  @keyup.enter="urlValid ? toggleConnection() : null"
-                />
-                <input
-                  id="socketio-path"
-                  v-model="path"
-                  class="flex flex-1 w-full px-4 py-2 border rounded-r bg-primaryLight border-divider text-secondaryDark"
-                  spellcheck="false"
-                  :disabled="connectionState"
-                  @keyup.enter="urlValid ? toggleConnection() : null"
-                />
-              </div>
-              <ButtonPrimary
-                id="connect"
-                :disabled="!urlValid"
-                name="connect"
-                class="w-32"
-                :label="
-                  !connectionState
-                    ? $t('action.connect')
-                    : $t('action.disconnect')
-                "
-                :loading="connectingState"
-                @click.native="toggleConnection"
-              />
-            </div>
-          </div>
-          <div
-            class="sticky z-10 flex items-center justify-between pl-4 border-b bg-primary border-dividerLight top-upperPrimaryStickyFold"
-          >
-            <span class="flex items-center">
-              <label class="font-semibold text-secondaryLight">
-                {{ $t("authorization.type") }}
-              </label>
+  <AppPaneLayout>
+    <template #primary>
+      <div
+        class="sticky top-0 z-10 flex flex-shrink-0 p-4 overflow-x-auto space-x-2 bg-primary hide-scrollbar"
+      >
+        <div class="inline-flex flex-1 space-x-2">
+          <div class="flex flex-1">
+            <label for="client-version">
               <tippy
-                ref="authTypeOptions"
+                ref="versionOptions"
                 interactive
                 trigger="click"
                 theme="popover"
@@ -105,132 +16,193 @@
               >
                 <template #trigger>
                   <span class="select-wrapper">
-                    <ButtonSecondary
-                      class="pr-8 ml-2 rounded-none"
-                      :label="authType"
+                    <input
+                      id="client-version"
+                      v-tippy="{ theme: 'tooltip' }"
+                      title="socket.io-client version"
+                      class="flex px-4 py-2 font-semibold border rounded-l cursor-pointer bg-primaryLight border-divider text-secondaryDark w-26"
+                      :value="`Client ${clientVersion}`"
+                      readonly
+                      :disabled="connectionState"
                     />
                   </span>
                 </template>
                 <div class="flex flex-col" role="menu">
                   <SmartItem
-                    label="None"
-                    :icon="
-                      authType === 'None'
-                        ? 'radio_button_checked'
-                        : 'radio_button_unchecked'
-                    "
-                    :active="authType === 'None'"
-                    @click.native="
-                      () => {
-                        authType = 'None'
-                        authTypeOptions.tippy().hide()
-                      }
-                    "
-                  />
-                  <SmartItem
-                    label="Bearer Token"
-                    :icon="
-                      authType === 'Bearer'
-                        ? 'radio_button_checked'
-                        : 'radio_button_unchecked'
-                    "
-                    :active="authType === 'Bearer'"
-                    @click.native="
-                      () => {
-                        authType = 'Bearer'
-                        authTypeOptions.tippy().hide()
-                      }
-                    "
+                    v-for="(_, version) in socketIoClients"
+                    :key="`client-${version}`"
+                    :label="`Client ${version}`"
+                    @click.native="onSelectVersion(version)"
                   />
                 </div>
               </tippy>
-            </span>
-            <div class="flex">
-              <SmartCheckbox
-                :on="authActive"
-                class="px-2"
-                @change="authActive = !authActive"
-              >
-                {{ $t("state.enabled") }}
-              </SmartCheckbox>
-              <ButtonSecondary
-                v-tippy="{ theme: 'tooltip' }"
-                to="https://docs.hoppscotch.io/features/authorization"
-                blank
-                :title="$t('app.wiki')"
-                svg="help-circle"
+            </label>
+            <input
+              id="socketio-url"
+              v-model="url"
+              type="url"
+              autocomplete="off"
+              spellcheck="false"
+              :class="{ error: !urlValid }"
+              class="flex flex-1 w-full px-4 py-2 border bg-primaryLight border-divider text-secondaryDark"
+              :placeholder="$t('socketio.url')"
+              :disabled="connectionState"
+              @keyup.enter="urlValid ? toggleConnection() : null"
+            />
+            <input
+              id="socketio-path"
+              v-model="path"
+              class="flex flex-1 w-full px-4 py-2 border rounded-r bg-primaryLight border-divider text-secondaryDark"
+              spellcheck="false"
+              :disabled="connectionState"
+              @keyup.enter="urlValid ? toggleConnection() : null"
+            />
+          </div>
+          <ButtonPrimary
+            id="connect"
+            :disabled="!urlValid"
+            name="connect"
+            class="w-32"
+            :label="
+              !connectionState ? $t('action.connect') : $t('action.disconnect')
+            "
+            :loading="connectingState"
+            @click.native="toggleConnection"
+          />
+        </div>
+      </div>
+      <div
+        class="sticky z-10 flex items-center justify-between pl-4 border-b bg-primary border-dividerLight top-upperPrimaryStickyFold"
+      >
+        <span class="flex items-center">
+          <label class="font-semibold text-secondaryLight">
+            {{ $t("authorization.type") }}
+          </label>
+          <tippy
+            ref="authTypeOptions"
+            interactive
+            trigger="click"
+            theme="popover"
+            arrow
+          >
+            <template #trigger>
+              <span class="select-wrapper">
+                <ButtonSecondary
+                  class="pr-8 ml-2 rounded-none"
+                  :label="authType"
+                />
+              </span>
+            </template>
+            <div class="flex flex-col" role="menu">
+              <SmartItem
+                label="None"
+                :icon="
+                  authType === 'None'
+                    ? 'radio_button_checked'
+                    : 'radio_button_unchecked'
+                "
+                :active="authType === 'None'"
+                @click.native="
+                  () => {
+                    authType = 'None'
+                    authTypeOptions.tippy().hide()
+                  }
+                "
               />
-              <ButtonSecondary
-                v-tippy="{ theme: 'tooltip' }"
-                :title="$t('action.clear')"
-                svg="trash-2"
-                @click.native="clearContent"
+              <SmartItem
+                label="Bearer Token"
+                :icon="
+                  authType === 'Bearer'
+                    ? 'radio_button_checked'
+                    : 'radio_button_unchecked'
+                "
+                :active="authType === 'Bearer'"
+                @click.native="
+                  () => {
+                    authType = 'Bearer'
+                    authTypeOptions.tippy().hide()
+                  }
+                "
               />
             </div>
-          </div>
-          <div
-            v-if="authType === 'None'"
-            class="flex flex-col items-center justify-center p-4 text-secondaryLight"
+          </tippy>
+        </span>
+        <div class="flex">
+          <SmartCheckbox
+            :on="authActive"
+            class="px-2"
+            @change="authActive = !authActive"
           >
-            <img
-              :src="`/images/states/${$colorMode.value}/login.svg`"
-              loading="lazy"
-              class="inline-flex flex-col object-contain object-center w-16 h-16 my-4"
-              :alt="$t('empty.authorization')"
-            />
-            <span class="pb-4 text-center">
-              This SocketIO connection does not use any authentication.
-            </span>
-            <ButtonSecondary
-              outline
-              :label="$t('app.documentation')"
+            {{ $t("state.enabled") }}
+          </SmartCheckbox>
+          <ButtonSecondary
+            v-tippy="{ theme: 'tooltip' }"
+            to="https://docs.hoppscotch.io/features/authorization"
+            blank
+            :title="$t('app.wiki')"
+            svg="help-circle"
+          />
+          <ButtonSecondary
+            v-tippy="{ theme: 'tooltip' }"
+            :title="$t('action.clear')"
+            svg="trash-2"
+            @click.native="clearContent"
+          />
+        </div>
+      </div>
+      <div
+        v-if="authType === 'None'"
+        class="flex flex-col items-center justify-center p-4 text-secondaryLight"
+      >
+        <img
+          :src="`/images/states/${$colorMode.value}/login.svg`"
+          loading="lazy"
+          class="inline-flex flex-col object-contain object-center w-16 h-16 my-4"
+          :alt="$t('empty.authorization')"
+        />
+        <span class="pb-4 text-center">
+          This SocketIO connection does not use any authentication.
+        </span>
+        <ButtonSecondary
+          outline
+          :label="$t('app.documentation')"
+          to="https://docs.hoppscotch.io/features/authorization"
+          blank
+          svg="external-link"
+          reverse
+          class="mb-4"
+        />
+      </div>
+      <div
+        v-if="authType === 'Bearer'"
+        class="flex flex-1 border-b border-dividerLight"
+      >
+        <div class="w-2/3 border-r border-dividerLight">
+          <div class="flex flex-1 border-b border-dividerLight">
+            <SmartEnvInput v-model="bearerToken" placeholder="Token" />
+          </div>
+        </div>
+        <div
+          class="sticky h-full p-4 overflow-auto bg-primary top-upperTertiaryStickyFold min-w-46 max-w-1/3 z-9"
+        >
+          <div class="p-2">
+            <div class="pb-2 text-secondaryLight">
+              {{ $t("helpers.authorization") }}
+            </div>
+            <SmartAnchor
+              class="link"
+              :label="`${$t('authorization.learn')} \xA0 →`"
               to="https://docs.hoppscotch.io/features/authorization"
               blank
-              svg="external-link"
-              reverse
-              class="mb-4"
             />
           </div>
-          <div
-            v-if="authType === 'Bearer'"
-            class="flex flex-1 border-b border-dividerLight"
-          >
-            <div class="w-2/3 border-r border-dividerLight">
-              <div class="flex flex-1 border-b border-dividerLight">
-                <SmartEnvInput v-model="bearerToken" placeholder="Token" />
-              </div>
-            </div>
-            <div
-              class="sticky h-full p-4 overflow-auto bg-primary top-upperTertiaryStickyFold min-w-46 max-w-1/3 z-9"
-            >
-              <div class="p-2">
-                <div class="pb-2 text-secondaryLight">
-                  {{ $t("helpers.authorization") }}
-                </div>
-                <SmartAnchor
-                  class="link"
-                  :label="`${$t('authorization.learn')} \xA0 →`"
-                  to="https://docs.hoppscotch.io/features/authorization"
-                  blank
-                />
-              </div>
-            </div>
-          </div>
-        </Pane>
-        <Pane
-          :size="COLUMN_LAYOUT ? 65 : 50"
-          class="hide-scrollbar !overflow-auto flex flex-col"
-        >
-          <RealtimeLog :title="$t('socketio.log')" :log="log" />
-        </Pane>
-      </Splitpanes>
-    </Pane>
-    <Pane
-      v-if="SIDEBAR"
-      size="25"
-      min-size="20"
-      class="hide-scrollbar !overflow-auto flex flex-col"
-    >
+        </div>
+      </div>
+    </template>
+    <template #secondary>
+      <RealtimeLog :title="$t('socketio.log')" :log="log" />
+    </template>
+    <template #sidebar>
       <div class="flex items-center justify-between p-4">
         <label for="events" class="font-semibold text-secondaryLight">
           {{ $t("socketio.events") }}
@@ -297,24 +269,19 @@
           </div>
         </div>
       </div>
-    </Pane>
-  </Splitpanes>
+    </template>
+  </AppPaneLayout>
 </template>
 
 <script>
 import { defineComponent, ref } from "@nuxtjs/composition-api"
-import { Splitpanes, Pane } from "splitpanes"
-import "splitpanes/dist/splitpanes.css"
 // All Socket.IO client version imports
 import ClientV2 from "socket.io-client-v2"
 import { io as ClientV3 } from "socket.io-client-v3"
 import { io as ClientV4 } from "socket.io-client-v4"
-
 import wildcard from "socketio-wildcard"
 import debounce from "lodash/debounce"
-import { breakpointsTailwind, useBreakpoints } from "@vueuse/core"
 import { logHoppRequestRunToAnalytics } from "~/helpers/fb/analytics"
-import { useSetting } from "~/newstore/settings"
 import {
   SIOEndpoint$,
   setSIOEndpoint,
@@ -341,16 +308,8 @@ const socketIoClients = {
 }
 
 export default defineComponent({
-  components: { Splitpanes, Pane },
   setup() {
-    const breakpoints = useBreakpoints(breakpointsTailwind)
-    const mdAndLarger = breakpoints.greater("md")
-
     return {
-      mdAndLarger,
-      SIDEBAR: useSetting("SIDEBAR"),
-      COLUMN_LAYOUT: useSetting("COLUMN_LAYOUT"),
-      SIDEBAR_ON_LEFT: useSetting("SIDEBAR_ON_LEFT"),
       socketIoClients,
       url: useStream(SIOEndpoint$, "", setSIOEndpoint),
       clientVersion: useStream(SIOVersion$, "", setSIOVersion),
